Split into federated and non-federated test groups
